requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\SearchUI.py: 9
Pillow == 7.1.2

# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\SearchUI.py: 5
PySimpleGUI == 4.34.0

# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\WebScraper.py: 3
bs4 == 0.0.1

# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\Indexer.py: 15
# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\Test.py: 15
h5py == 2.10.0

# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\WebScraper.py: 1
lxml == 4.6.2

# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\Indexer.py: 4,6,7,8,9
# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\Test.py: 4,6,7,8,9
nltk == 3.5

# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\Indexer.py: 5
# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\Test.py: 5
numpy == 1.19.1

# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\Indexer.py: 13
# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\Test.py: 13
pandas == 1.0.3

# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\SearchUI.py: 11
# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\WebScraper.py: 2
requests == 2.23.0

# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\Indexer.py: 16,17
# C:\Users\srina\PycharmProjects\Steam_GameRetrieval\Test.py: 16,17
scipy == 1.5.0
